Consolidate Checkout shipping scenario
Committed-By-Agent: codex
