+++ paymentsheet-example/src/main/AndroidManifest.xml
@@ -61,9 +61,6 @@
             android:name="com.stripe.android.paymentsheet.example.playground.checkout.CheckoutControllerExampleActivity"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name="com.stripe.android.paymentsheet.example.playground.checkout.CheckoutShippingAddressElementPlaygroundActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity
             android:name="com.stripe.android.paymentsheet.example.playground.PaymentSheetPlaygroundActivity"
             android:theme="@style/AppTheme.NoActionBar"
             android:exported="true"
