+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/MainActivity.kt
@@ -34,7 +34,6 @@ import com.stripe.android.core.version.StripeSdkVersion
 import com.stripe.android.paymentsheet.example.databinding.ActivityMainBinding
 import com.stripe.android.paymentsheet.example.playground.PaymentSheetPlaygroundActivity
 import com.stripe.android.paymentsheet.example.playground.checkout.CheckoutControllerExampleActivity
-import com.stripe.android.paymentsheet.example.playground.checkout.CheckoutShippingAddressElementPlaygroundActivity
 import com.stripe.android.paymentsheet.example.playground.embedded.EmbeddedExampleActivity
 import com.stripe.android.paymentsheet.example.samples.ui.SECTION_ALPHA
 import com.stripe.android.paymentsheet.example.samples.ui.addresselement.AddressElementExampleActivity
@@ -66,12 +65,6 @@ class MainActivity : AppCompatActivity() {
                 titleResId = R.string.checkout_controller_example_title,
                 subtitleResId = R.string.checkout_controller_example_subtitle,
                 klass = CheckoutControllerExampleActivity::class.java,
-                section = MenuItem.Section.Internal,
-            ),
-            MenuItem(
-                titleResId = R.string.checkout_shipping_address_element_title,
-                subtitleResId = R.string.checkout_shipping_address_element_subtitle,
-                klass = CheckoutShippingAddressElementPlaygroundActivity::class.java,
                 section = MenuItem.Section.Internal,
             ),
             MenuItem(
