+++ paymentsheet-example/src/main/res/values/strings.xml
@@ -20,8 +20,6 @@
     <string name="playground_subtitle">Testing playground for Stripe engineers</string>
     <string name="checkout_controller_example_title">CheckoutController Example</string>
     <string name="checkout_controller_example_subtitle">Simple CheckoutController integration</string>
-    <string name="checkout_shipping_address_element_title">Checkout Shipping Address Element</string>
-    <string name="checkout_shipping_address_element_subtitle">Manually confirm and verify the refreshed shipping address</string>
     <string name="onramp_title">Onramp Coordinator</string>
     <string name="onramp_subtitle">Test the OnrampCoordinator functionality</string>
     <string name="action_settings">Settings</string>
